compatibility-tests/sdk-test-appconfig/requirements.txt
pytest==8.3.3
azure-appconfiguration==1.7.1
